Finished global verification for found civ

+++ Interfaz/interfaz/mainwindow.cpp
@@ -166,11 +166,14 @@ void MainWindow::on_search_civilization_input_clicked()
     QString search = ui->civilization_search_input->text();
 
     if (videogame.total() == 0) {
+        flag = false;
         QMessageBox::information(this, "Error", "There are no civilizations to search");
     } else if (!videogame.civilizationExists(search.toStdString())) {
+        flag = false;
         QMessageBox::information(this, "Error", "Civilization does not exist");
     } else {
         qDebug() << "[✔] Civilization found";
+        flag = true;
     }
 }
 
@@ -179,23 +182,25 @@ void MainWindow::on_search_civilization_input_clicked()
  */
 void MainWindow::on_push_villager_front_clicked()
 {
-    Villager v;
-    QString search = ui->civilization_search_input->text();
-    Civilization* c = videogame.searchCivilization(search.toStdString());
-    QString name = ui->villager_name_add->text();
-    int age = ui->villager_age_add->value();
-    QString gender = ui->villager_gender_add->text();
-    int health = ui->villager_health_add->value();
+    if (flag) {
+        Villager v;
+        QString search = ui->civilization_search_input->text();
+        Civilization* c = videogame.searchCivilization(search.toStdString());
+        QString name = ui->villager_name_add->text();
+        int age = ui->villager_age_add->value();
+        QString gender = ui->villager_gender_add->text();
+        int health = ui->villager_health_add->value();
 
-    v.setNombre(name.toStdString());
-    v.setEdad(age);
-    v.setGender(gender.toStdString());
-    v.setHealth(health);
+        v.setNombre(name.toStdString());
+        v.setEdad(age);
+        v.setGender(gender.toStdString());
+        v.setHealth(health);
 
-    c->push_villager_front(v);
-    qDebug() << "[✔] Villager pushed to front";
-
-
+        c->push_villager_front(v);
+        qDebug() << "[✔] Villager pushed to front";
+    } else {
+        QMessageBox::information(this, "Error", "Civilization not found");
+    }
 }
 
 /**
@@ -203,21 +208,26 @@ void MainWindow::on_push_villager_front_clicked()
  */
 void MainWindow::on_push_villager_back_clicked()
 {
-    Villager v;
-    QString search = ui->civilization_search_input->text();
-    Civilization* c = videogame.searchCivilization(search.toStdString());
-    QString name = ui->villager_name_add->text();
-    int age = ui->villager_age_add->value();
-    QString gender = ui->villager_gender_add->text();
-    int health = ui->villager_health_add->value();
+    if (flag) {
+        Villager v;
+        QString search = ui->civilization_search_input->text();
+        Civilization* c = videogame.searchCivilization(search.toStdString());
+        QString name = ui->villager_name_add->text();
+        int age = ui->villager_age_add->value();
+        QString gender = ui->villager_gender_add->text();
+        int health = ui->villager_health_add->value();
 
-    v.setNombre(name.toStdString());
-    v.setEdad(age);
-    v.setGender(gender.toStdString());
-    v.setHealth(health);
+        v.setNombre(name.toStdString());
+        v.setEdad(age);
+        v.setGender(gender.toStdString());
+        v.setHealth(health);
 
-    c->push_villager_back(v);
-    qDebug() << "[✔] Villager pushed to back";
+        c->push_villager_back(v);
+        qDebug() << "[✔] Villager pushed to back";
+
+    } else {
+        QMessageBox::information(this, "Error", "Civilization not found");
+    }
 }
 
 /**
@@ -225,17 +235,21 @@ void MainWindow::on_push_villager_back_clicked()
  */
 void MainWindow::on_delete_villager_name_submit_clicked()
 {
-    QString search = ui->civilization_search_input->text();
-    Civilization* c = videogame.searchCivilization(search.toStdString());
-    QString name = ui->villager_delete_name_input->text();
+    if (flag) {
+        QString search = ui->civilization_search_input->text();
+        Civilization* c = videogame.searchCivilization(search.toStdString());
+        QString name = ui->villager_delete_name_input->text();
 
-    if (c->villager_size() == 0) {
-        QMessageBox::information(this, "Error", "There are no villagers to delete");
-    } else if (!c->delete_villager(name.toStdString())) {
-        QMessageBox::information(this, "Error", "Villager does not exist");
-    } else {
-        qDebug() << "[✔]" << "Unfortunate villager deleted";
-    }
+        if (c->villager_size() == 0) {
+            QMessageBox::information(this, "Error", "There are no villagers to delete");
+        } else if (!c->delete_villager(name.toStdString())) {
+            QMessageBox::information(this, "Error", "Villager does not exist");
+        } else {
+            qDebug() << "[✔]" << "Unfortunate villager deleted";
+        }
+
+    } else
+        QMessageBox::information(this, "Error", "Civilization not found");
 
 }
 
@@ -244,66 +258,105 @@ void MainWindow::on_delete_villager_name_submit_clicked()
  */
 void MainWindow::on_delete_villager_health_submit_clicked()
 {
-    QString search = ui->civilization_search_input->text();
-    int health = ui->delete_villager_health_input->value();
-    Civilization* c = videogame.searchCivilization(search.toStdString());
+    if (flag) {
+        QString search = ui->civilization_search_input->text();
+        int health = ui->delete_villager_health_input->value();
+        Civilization* c = videogame.searchCivilization(search.toStdString());
 
-    if (c->villager_size() == 0) {
-        QMessageBox::information(this, "Error", "There are no villagers to delete");
-    } else if (!c->delete_villagers(health)) {
-        QMessageBox::information(this, "Error", "There are no villagers below this health");
-    } else {
-        qDebug() << "[✔]" << "Sick villager deleted";
-    }
+        if (c->villager_size() == 0) {
+            QMessageBox::information(this, "Error", "There are no villagers to delete");
+        } else if (!c->delete_villagers(health)) {
+            QMessageBox::information(this, "Error", "There are no villagers below this health");
+        } else {
+            qDebug() << "[✔]" << "Sick villager deleted";
+        }
+
+    } else
+        QMessageBox::information(this, "Error", "Civilization not found");
 
 }
-
+/**
+ * Delete old villagers
+ */
 void MainWindow::on_delete_old_villagers_clicked()
 {
-    QString search = ui->civilization_search_input->text();
-    Civilization* c = videogame.searchCivilization(search.toStdString());
+    if (flag) {
+        QString search = ui->civilization_search_input->text();
+        Civilization* c = videogame.searchCivilization(search.toStdString());
 
-    if (c->villager_size() == 0) {
-        QMessageBox::information(this, "Error", "There are no villagers to delete");
-    } else if (!c->delete_villager_age()) {
-        QMessageBox::information(this, "Error", "There are no old villagers to delete");
-    } else {
-        qDebug() << "[✔]" << "Old villager deleted";
-    }
+        if (c->villager_size() == 0) {
+            QMessageBox::information(this, "Error", "There are no villagers to delete");
+        } else if (!c->delete_villager_age()) {
+            QMessageBox::information(this, "Error", "There are no old villagers to delete");
+        } else {
+            qDebug() << "[✔]" << "Old villager deleted";
+        }
+
+    } else
+        QMessageBox::information(this, "Error", "Civilization not found");
 }
 
+/**
+ * Order villagers by name
+ */
 void MainWindow::on_order_villagers_name_submit_clicked()
 {
-    QString search = ui->civilization_search_input->text();
-    Civilization* c = videogame.searchCivilization(search.toStdString());
+    if (flag) {
+        QString search = ui->civilization_search_input->text();
+        Civilization* c = videogame.searchCivilization(search.toStdString());
 
-    c->sort_villagers_name();
-    qDebug() << "[✔]" << "Villagers sorted by name";
+        c->sort_villagers_name();
+        qDebug() << "[✔]" << "Villagers sorted by name";
+
+    } else
+        QMessageBox::information(this, "Error", "Civilization not found");
 }
 
+/**
+ * Order villagers by age
+ */
 void MainWindow::on_order_villagers_age_submit_clicked()
 {
-    QString search = ui->civilization_search_input->text();
-    Civilization* c = videogame.searchCivilization(search.toStdString());
+    if (flag) {
+        QString search = ui->civilization_search_input->text();
+        Civilization* c = videogame.searchCivilization(search.toStdString());
 
-    c->sort_villagers_age();
-    qDebug() << "[✔]" << "Villagers sorted by age";
+        c->sort_villagers_age();
+        qDebug() << "[✔]" << "Villagers sorted by age";
+
+    } else
+        QMessageBox::information(this, "Error", "Civilization not found");
 }
 
+/**
+ * Order villagers by health
+ */
 void MainWindow::on_order_villagers_by_health_clicked()
 {
-    QString search = ui->civilization_search_input->text();
-    Civilization* c = videogame.searchCivilization(search.toStdString());
+    if (flag) {
+        QString search = ui->civilization_search_input->text();
+        Civilization* c = videogame.searchCivilization(search.toStdString());
 
-    c->sort_villagers_health();
-    qDebug() << "[✔]" << "Villagers sorted by health";
+        c->sort_villagers_health();
+        qDebug() << "[✔]" << "Villagers sorted by health";
+
+    } else
+        QMessageBox::information(this, "Error", "Civilization not found");
 }
 
+/**
+ * Show villagers
+ */
 void MainWindow::on_show_villagers_input_clicked()
 {
-    QString search = ui->civilization_search_input->text();
-    Civilization* c = videogame.searchCivilization(search.toStdString());
+    if (flag) {
+        QString search = ui->civilization_search_input->text();
+        Civilization* c = videogame.searchCivilization(search.toStdString());
 
-    villagerDisplay.show();
-    villagerDisplay.setCivilization(c);
+        villagerDisplay.show();
+        villagerDisplay.setCivilization(c);
+        qDebug() << "[✔] Showing villagers";
+
+    } else
+        QMessageBox::information(this, "Error", "Civilization not found");
 }
